Split {Base,Triangle,Tetra}Mesh from mesh_container.py into mesh.py
