+++ nanomesh/__init__.py
@@ -1,9 +1,10 @@
 import logging
 import sys
 
+from .mesh import LineMesh, TetraMesh, TriangleMesh
 from .mesh2d import Mesher2D, compare_mesh_with_image, simple_triangulate
 from .mesh3d import Mesher3D
-from .mesh_container import TetraMesh, TriangleMesh
+from .mesh_container import MeshContainer
 from .plane import Plane
 from .volume import Volume
 
@@ -20,8 +21,10 @@ __all__ = [
     '__email__',
     '__version__',
     'compare_mesh_with_image',
+    'LineMesh',
     'Mesher2D',
     'Mesher3D',
+    'MeshContainer',
     'Plane',
     'simple_triangulate',
     'TetraMesh',
