+++ nanomesh/metrics.py
@@ -4,7 +4,7 @@ from typing import Callable, Optional, Tuple
 import matplotlib.pyplot as plt
 import numpy as np
 
-from .mesh_container import BaseMesh
+from .mesh import BaseMesh
 
 
 class Metric:
