+++ tests/test_plane.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 from matplotlib.testing.decorators import image_comparison
 
-from nanomesh.mesh_container import TriangleMesh
+from nanomesh.mesh import TriangleMesh
 from nanomesh.plane import Plane
